+++ Tests/Wire/Resources/Users/UsersClientWireTests.swift
@@ -44,6 +44,13 @@ import Chrt
                   "stripe_customer_id": "stripe_customer_id",
                   "primary_email_address": "primary_email_address",
                   "phone_number": "phone_number",
+                  "firebase_cloud_messaging_token_data": [
+                    {
+                      "firebase_cloud_messaging_token": "firebase_cloud_messaging_token",
+                      "platform": "android",
+                      "last_used_timestamp": "2024-01-15T09:30:00Z"
+                    }
+                  ],
                   "_id": "_id"
                 }
                 """.utf8
@@ -61,6 +68,13 @@ import Chrt
             stripeCustomerId: Optional("stripe_customer_id"),
             primaryEmailAddress: Optional("primary_email_address"),
             phoneNumber: Optional("phone_number"),
+            firebaseCloudMessagingTokenData: Optional([
+                FirebaseCloudMessagingTokenData1(
+                    firebaseCloudMessagingToken: "firebase_cloud_messaging_token",
+                    platform: Optional(.android),
+                    lastUsedTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)
+                )
+            ]),
             id: "_id"
         )
         let response = try await client.users.getUserPrivateDataV1(requestOptions: RequestOptions(additionalHeaders: stub.headers))
@@ -79,6 +93,13 @@ import Chrt
                   "stripe_customer_id": "stripe_customer_id",
                   "primary_email_address": "primary_email_address",
                   "phone_number": "phone_number",
+                  "firebase_cloud_messaging_token_data": [
+                    {
+                      "firebase_cloud_messaging_token": "firebase_cloud_messaging_token",
+                      "platform": "android",
+                      "last_used_timestamp": "2024-01-15T09:30:00Z"
+                    }
+                  ],
                   "_id": "_id"
                 }
                 """.utf8
@@ -96,6 +117,13 @@ import Chrt
             stripeCustomerId: Optional("stripe_customer_id"),
             primaryEmailAddress: Optional("primary_email_address"),
             phoneNumber: Optional("phone_number"),
+            firebaseCloudMessagingTokenData: Optional([
+                FirebaseCloudMessagingTokenData1(
+                    firebaseCloudMessagingToken: "firebase_cloud_messaging_token",
+                    platform: Optional(.android),
+                    lastUsedTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)
+                )
+            ]),
             id: "_id"
         )
         let response = try await client.users.createUserPrivateDataV1(
@@ -117,6 +145,13 @@ import Chrt
                   "stripe_customer_id": "stripe_customer_id",
                   "primary_email_address": "primary_email_address",
                   "phone_number": "phone_number",
+                  "firebase_cloud_messaging_token_data": [
+                    {
+                      "firebase_cloud_messaging_token": "firebase_cloud_messaging_token",
+                      "platform": "android",
+                      "last_used_timestamp": "2024-01-15T09:30:00Z"
+                    }
+                  ],
                   "_id": "_id"
                 }
                 """.utf8
@@ -134,10 +169,69 @@ import Chrt
             stripeCustomerId: Optional("stripe_customer_id"),
             primaryEmailAddress: Optional("primary_email_address"),
             phoneNumber: Optional("phone_number"),
+            firebaseCloudMessagingTokenData: Optional([
+                FirebaseCloudMessagingTokenData1(
+                    firebaseCloudMessagingToken: "firebase_cloud_messaging_token",
+                    platform: Optional(.android),
+                    lastUsedTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)
+                )
+            ]),
             id: "_id"
         )
         let response = try await client.users.updateUserPrivateDataV1(
             request: .init(),
+            requestOptions: RequestOptions(additionalHeaders: stub.headers)
+        )
+        try #require(response == expectedResponse)
+    }
+
+    @Test func upsertFirebaseCloudMessagingTokenV11() async throws -> Void {
+        let stub = HTTPStub()
+        stub.setResponse(
+            body: Data(
+                """
+                {
+                  "schema_version": 1,
+                  "created_at_timestamp": "2024-01-15T09:30:00Z",
+                  "user_id": "user_id",
+                  "stripe_customer_id": "stripe_customer_id",
+                  "primary_email_address": "primary_email_address",
+                  "phone_number": "phone_number",
+                  "firebase_cloud_messaging_token_data": [
+                    {
+                      "firebase_cloud_messaging_token": "firebase_cloud_messaging_token",
+                      "platform": "android",
+                      "last_used_timestamp": "2024-01-15T09:30:00Z"
+                    }
+                  ],
+                  "_id": "_id"
+                }
+                """.utf8
+            )
+        )
+        let client = ChrtClient(
+            baseURL: "https://api.fern.com",
+            token: "<token>",
+            urlSession: stub.urlSession
+        )
+        let expectedResponse = UserPrivateData1(
+            schemaVersion: 1,
+            createdAtTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
+            userId: "user_id",
+            stripeCustomerId: Optional("stripe_customer_id"),
+            primaryEmailAddress: Optional("primary_email_address"),
+            phoneNumber: Optional("phone_number"),
+            firebaseCloudMessagingTokenData: Optional([
+                FirebaseCloudMessagingTokenData1(
+                    firebaseCloudMessagingToken: "firebase_cloud_messaging_token",
+                    platform: Optional(.android),
+                    lastUsedTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)
+                )
+            ]),
+            id: "_id"
+        )
+        let response = try await client.users.upsertFirebaseCloudMessagingTokenV1(
+            request: .init(firebaseCloudMessagingToken: "firebase_cloud_messaging_token"),
             requestOptions: RequestOptions(additionalHeaders: stub.headers)
         )
         try #require(response == expectedResponse)